Add unit tests (#14)
* Move IProdcutService from namespace

* Make product price a double

* Add unit tests

+++ GraphQL.Sample/GraphQL.Sample.Infra.Data/Migrations/20201208002510_InitialMigration.cs
@@ -15,7 +15,7 @@ namespace GraphQL.Sample.Infra.Data.Migrations
                     Name = table.Column<string>(nullable: true),
                     Category = table.Column<int>(nullable: false),
                     Description = table.Column<string>(nullable: true),
-                    Price = table.Column<decimal>(nullable: false)
+                    Price = table.Column<double>(nullable: false)
                 },
                 constraints: table =>
                 {
